Major refactor of tests
This large commit refactors the way tests are run from targeting
specially committed conf files for testing to the same conf files as
those which will be deployed in production. See the README for a
description of some of the limitations of this. A nice side-effect is
that the migration script is a lot cleaner now.